Documents/CLAUDE/[deep]frame/liminal-board/server/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.0
python-socketio==5.11.0
python-multipart==0.0.6
pydantic==2.6.1
openai==1.12.0